refactor: move view/util into just util
