+++ core/src/main/java/dk/gtz/graphedit/tool/AbstractBaseTool.java
@@ -2,10 +2,10 @@ package dk.gtz.graphedit.tool;
 
 import java.util.Optional;
 
-import dk.gtz.graphedit.view.events.EdgeMouseEvent;
-import dk.gtz.graphedit.view.events.VertexMouseEvent;
-import dk.gtz.graphedit.view.events.ViewportKeyEvent;
-import dk.gtz.graphedit.view.events.ViewportMouseEvent;
+import dk.gtz.graphedit.events.EdgeMouseEvent;
+import dk.gtz.graphedit.events.VertexMouseEvent;
+import dk.gtz.graphedit.events.ViewportKeyEvent;
+import dk.gtz.graphedit.events.ViewportMouseEvent;
 
 public abstract class AbstractBaseTool implements ITool {
     @Override
