+++ core/src/main/java/dk/gtz/graphedit/tool/VertexCreateTool.java
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.model.ModelVertex;
 import dk.gtz.graphedit.view.ISyntaxFactory;
-import dk.gtz.graphedit.view.events.ViewportMouseEvent;
+import dk.gtz.graphedit.events.ViewportMouseEvent;
 import dk.gtz.graphedit.viewmodel.ViewModelGraph;
 import dk.gtz.graphedit.viewmodel.ViewModelPoint;
 import dk.yalibs.yadi.DI;
