+++ core/src/main/java/dk/gtz/graphedit/util/HeightDragResizer.java
@@ -1,4 +1,4 @@
-package dk.gtz.graphedit.view.util;
+package dk.gtz.graphedit.util;
 
 import javafx.scene.Cursor;
 import javafx.scene.input.MouseEvent;
