+++ core/src/main/java/dk/gtz/graphedit/util/IObservableUndoSystem.java
@@ -1,4 +1,4 @@
-package dk.gtz.graphedit.view.util;
+package dk.gtz.graphedit.util;
 
 import dk.yalibs.yaundo.IUndoSystem;
 import dk.yalibs.yaundo.Undoable;
