+++ core/src/main/java/dk/gtz/graphedit/view/EdgeController.java
@@ -6,8 +6,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.tool.ITool;
-import dk.gtz.graphedit.view.events.EdgeMouseEvent;
-import dk.gtz.graphedit.view.util.BindingsUtil;
+import dk.gtz.graphedit.events.EdgeMouseEvent;
+import dk.gtz.graphedit.util.BindingsUtil;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelPoint;
