+++ core/src/main/java/dk/gtz/graphedit/view/ProjectSettingsController.java
@@ -8,7 +8,7 @@ import org.slf4j.LoggerFactory;
 import atlantafx.base.controls.Tile;
 import atlantafx.base.theme.Styles;
 import dk.gtz.graphedit.tool.EditorActions;
-import dk.gtz.graphedit.view.util.InspectorUtils;
+import dk.gtz.graphedit.util.InspectorUtils;
 import dk.gtz.graphedit.viewmodel.ViewModelProject;
 import dk.yalibs.yadi.DI;
 import javafx.beans.property.ListProperty;
