+++ graphedit/src/main/java/dk/gtz/graphedit/view/GraphEditApplication.java
@@ -28,8 +28,8 @@ import dk.gtz.graphedit.tool.VertexCreateTool;
 import dk.gtz.graphedit.tool.VertexDeleteTool;
 import dk.gtz.graphedit.tool.VertexDragMoveTool;
 import dk.gtz.graphedit.tool.ViewTool;
-import dk.gtz.graphedit.view.util.IObservableUndoSystem;
-import dk.gtz.graphedit.view.util.ObservableStackUndoSystem;
+import dk.gtz.graphedit.util.IObservableUndoSystem;
+import dk.gtz.graphedit.util.ObservableStackUndoSystem;
 import dk.gtz.graphedit.viewmodel.FileBufferContainer;
 import dk.gtz.graphedit.viewmodel.IBufferContainer;
 import dk.gtz.graphedit.viewmodel.ISelectable;
