+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/view/ArcController.java
@@ -9,7 +9,7 @@ import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelArc;
 import dk.gtz.graphedit.tool.ITool;
 import dk.gtz.graphedit.view.EdgeController;
 import dk.gtz.graphedit.view.ISyntaxFactory;
-import dk.gtz.graphedit.view.util.BindingsUtil;
+import dk.gtz.graphedit.util.BindingsUtil;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelProjectResource;
 import javafx.beans.property.DoubleProperty;
