+++ std/src/main/java/dk/gtz/graphedit/plugins/view/ProjectFilesViewController.java
@@ -27,9 +27,7 @@ import dk.gtz.graphedit.model.ModelVertex;
 import dk.gtz.graphedit.serialization.IMimeTypeChecker;
 import dk.gtz.graphedit.serialization.IModelSerializer;
 import dk.gtz.graphedit.tool.EditorActions;
-import dk.gtz.graphedit.view.util.GlobFileMatcher;
-import dk.gtz.graphedit.view.util.IconUtils;
-import dk.gtz.graphedit.view.util.PlatformUtils;
+import dk.gtz.graphedit.util.PlatformUtils;
 import dk.gtz.graphedit.viewmodel.IBufferContainer;
 import dk.gtz.graphedit.viewmodel.ViewModelProject;
 import dk.yalibs.yadi.DI;
